LPS-95260 Avoid duplicate code
